fix: add commands variants
