+++ applications/tari_base_node/src/commands/mod.rs
@@ -1,7 +1,6 @@
 mod args;
 pub mod cli;
 pub mod command;
-pub mod command_handler;
 pub mod parser;
 pub mod reader;
 pub mod status_line;
